material ui fix and rename component
