+++ client/src/components/admin/Admin.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import Layout from '../layout/Layout';
 import { Route, Redirect, Switch } from 'react-router-dom';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Dashboard from './dashboard/Dashboard';
+import DashboardPage from './dashboard/DashboardPage';
 import Properties from './properties/Properties';
 import WorkOrders from './workorders/WorkOrders';
 import SettingsPage from './settings/SettingsPage';
 import Billing from './billing/Billing';
 import NewTenant from './tenants/NewTenants';
+import Dashboard from '@material-ui/icons/Dashboard';
 import Business from '@material-ui/icons/Business';
 import Assignment from '@material-ui/icons/Assignment';
 import People from '@material-ui/icons/People';
@@ -16,6 +16,7 @@ import CreditCard from '@material-ui/icons/CreditCard';
 import { AuthUserContext } from '../session';
 
 const links = [
+  { name: 'Dashboard', url: '', icon: <Dashboard /> },
   { name: 'Properties', url: 'properties', icon: <Business /> },
   { name: 'Work Orders', url: 'work-orders', icon: <Assignment /> },
   { name: 'Tenants', url: 'tenants', icon: <People /> },
@@ -30,9 +31,8 @@ class Admin extends Component {
         {authUser =>
           authUser ? (
             <Layout links={links}>
-              <CssBaseline />
               <Switch>
-                <Route exact path="/admin" component={Dashboard} />
+                <Route exact path="/admin" component={DashboardPage} />
                 <Route exact path="/admin/properties" component={Properties} />
                 <Route exact path="/admin/work-orders" component={WorkOrders} />
                 <Route exact path="/admin/tenants" component={NewTenant} />
